Decouple a test client from NoopClient
